refactor(main): move everything needed by main in pkg
This is a basic refactor, which doesn't yet cleanup the code. In future
PRs, I'll remove actions-only code from here.
