+++ internal/config/defaults.go
@@ -1,7 +1,7 @@
 package config
 
 import (
-	"github.com/nobe4/gh-ln/internal/log"
+	"github.com/nobe4/gh-ln/pkg/log"
 )
 
 type RawDefaults struct {
