+++ internal/config/links_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	fmock "github.com/nobe4/gh-ln/internal/format/mock"
-	"github.com/nobe4/gh-ln/internal/github"
-	gmock "github.com/nobe4/gh-ln/internal/github/mock"
+	"github.com/nobe4/gh-ln/pkg/github"
+	gmock "github.com/nobe4/gh-ln/pkg/github/mock"
 )
 
 func TestCombineLinks(t *testing.T) {
