+++ internal/config/repo.go
@@ -3,8 +3,8 @@ package config
 import (
 	"strings"
 
-	"github.com/nobe4/gh-ln/internal/github"
-	"github.com/nobe4/gh-ln/internal/log"
+	"github.com/nobe4/gh-ln/pkg/github"
+	"github.com/nobe4/gh-ln/pkg/log"
 )
 
 const repoPartsCount = 2 // owner/repo
