+++ pkg/environment/environment.go
@@ -14,8 +14,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/nobe4/gh-ln/internal/github"
-	"github.com/nobe4/gh-ln/internal/log"
+	"github.com/nobe4/gh-ln/pkg/github"
+	"github.com/nobe4/gh-ln/pkg/log"
 )
 
 var (
