+++ pkg/github/auth_test.go
