+++ pkg/ln/process.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/nobe4/gh-ln/internal/config"
 	"github.com/nobe4/gh-ln/internal/format"
-	"github.com/nobe4/gh-ln/internal/github"
-	"github.com/nobe4/gh-ln/internal/log"
+	"github.com/nobe4/gh-ln/pkg/github"
+	"github.com/nobe4/gh-ln/pkg/log"
 )
 
 const (
